+++ changelog.md
@@ -1,5 +1,7 @@
 # 1.0.1
 - Added settings
-- 
+- refactored
+- changed logo to look slightly better
+- fixed being able to scroll while tabbed out
 # 1.0.0
 - Scrolling works with geode scrollbar and gd scrolls